+++ src/main/java/com/dghysc/hy/user/repo/UserRepository.java
@@ -3,7 +3,6 @@ package com.dghysc.hy.user.repo;
 import com.dghysc.hy.user.model.User;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 /**
@@ -12,7 +11,7 @@ import java.util.Optional;
  * @author lin864464995@163.com
  * @see org.springframework.data.repository.CrudRepository
  */
-public interface UserRepository extends CrudRepository<User, BigInteger> {
+public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findByUsername(String username);
 
     boolean existsByUsername(String username);
